Worked on updating and renaming pages, removed the multiple plotting pages in favor of new one, implementation isn't CLOSE to being done
