+++ src/app/gui.cpp
@@ -13,11 +13,10 @@
 #include "app/assets/texture.hpp"
 #include "app/context.hpp"
 #include "app/pages/home.hpp"
-#include "app/pages/rpm.hpp"
+#include "app/pages/plot.hpp"
 #include "app/pages/serialmon.hpp"
-#include "app/pages/shock.hpp"
 #include "app/pages/utils.hpp"
-#include "app/pages/view.hpp"
+#include "app/pages/analysis.hpp"
 #include "app/style.hpp"
 #include "core/time.hpp"
 
@@ -33,21 +32,20 @@ namespace mbr {
         }
     }
 
-namespace {
+
     [[nodiscard]] pages::page* create_page(page_type_t type, const std::shared_ptr<app_context>& ctx, QWidget* parent) {
         switch (type) {
         case page_type_t::HOME:   return new pages::home_page(ctx, parent);
-        case page_type_t::PLOT:    return new pages::rpm_page(ctx, parent);
-        case page_type_t::ANALYSIS:  return new pages::shock_page(ctx, parent);
+        case page_type_t::PLOT:    return new pages::plot_page(ctx, parent);
+        case page_type_t::ANALYSIS:  return new pages::analysis_page(ctx, parent);
         case page_type_t::SERIAL: return new pages::serial_page(ctx, parent);
         default:                  return nullptr;
         }
     }
 
-} // namespace
 
     void gui_t::change_page(page_type_t type) {
-        pages_->setCurrentWidget(type);
+        pages_->setCurrentWidget(page_lookup_[type]);
     }
