+++ src/app/pages/analysis.cpp
